fix(is-access-granted): Moved utilities functions to separate utilities file
